requirements.txt
manim
manim-voiceover
kokoro-manim-voiceover
openai
numpy
python-dotenv
kokoro